Rip out STM/watcher-based store implementation.
Return to the original direct-to-DB implementation, where store
functions all write directly to the backing database. To handle event
notification (for the indexer), we use a clojure.core.async channel
that is `mult`, which the indexer `tap`s into; this is more
lightweight, as we don't gather garbage -- the previous implementation
would gather committed/cancelled transactions in the store-state ref,
and we hadn't yet implemented a GC or reaper task for it.

YAGNI, and decomplect things instead, leaving concurrency semantics
(mostly) to the backing database. We will still need to implement some
handling for store transaction conflicts, however.