Ensure that conflicting transactions are returned to callers.
Ensure that a conflicting transaction does not proceed, and that the
system does not set its state to committed. To enforce this, at commit
time, check that the transaction can proceed and that the requested
document changes are all applicable (i.e. expected document version is
what is the current document version). If some or all of the changes
are not applicable, mark the transaction as in conflict -- it is the
caller's decision whether to retry the changes made in the transaction
or to drop the transaction altogether.