Reduce log chattiness when fetching missing or unknown documents.
Reduce the chattiness of the log when fetching missing or unknown
documents by changing the log level of the message to `debug`. This
makes it visible when debugging the system, but relatively silent in
normal operations. It was a mistake to have set it to `warn` in the
first place.