Include type when fetching document by name.
As we specified that names are unique to each type (so documents can
have the same name but differ in type), we need to include the type
when fetching by name. This was a mistake on my part, as originally
names were unique to each context, and not constrained to each type.