Split up indexer into multiple threads.
Have effectively multiple threads executing indexer tasks, and don't
trigger flushing, as we are already handling that inherently in the
pull architecture.

To accomodate the fact that the indexer tasks are now separate
threads, we also simplify how documents are queued to be indexed.

Finally, we set the indexer period to 5 seconds.