Don't munge document keys.
We were inadvertently munging document keys when reading from the
store, transforming them to kebab-case. Since the document attribute
contains user data, we should avoid performing any transformation on
it that is not reversible.

Since the only way to reverse this particular transformation using
camel-snake-kebab is to note the original casing of keys (e.g. whether
PascalCase, camelCase, etc.), and since the transformation of document
keys has no benefit to us, then don't do it at all for the
document. Store the document as fetched from Postgres, and reinject
that into the resulting map post-transform.