Also handle JSON keys that can't be coerced to keywords.
Since it is possible to have a JSON key that is not a valid Clojure
keyword (e.g. "foo bar"), and since `keyword` doesn't do any
validation[1] on its input, generating technically invalid keywords,
we need to do the validation ourselves and prevent the coercion of
JSON keys that are invalid Clojure keywords.

---
[1] https://groups.google.com/forum/#!topic/clojure/WvXYkvLoQhI/discussion