Track state as part of document version.
Previously, the document state was separate from the contents of the
document; it would have been theoretically possible to update the
state outside of the document. However, the way we were actually
handling state updates was similar enough to how we track document
data updates, but without the attendant benefits of actually tracking
state changes across versions.

So, rationalize document state and make it a part of the document
version row. Similarly, move the update timestamp to be part of the
document version row, so that fetching a particular version of the
document also fetches when that document version was created.