Handle multiple writes to same document in single transaction.
Handle the case where a client writes to the same document in the same
transaction. Previously, we didn't handle this well, as our SQL for
updating the document header to point to the proper version was
slightly incorrect, and didn't handle this particular case -- it
assumed that the immediate predecessor version was the expected
version. Fix it so that we instead track the expected version, and
update the new version accordingly, so that we effectively coalesce
the chain of writes.

We also fix the query on the conflict check, and the SQL lookup for
the current version of a document -- that one in particular used an
unnecessary recursive CTE to get the version list; we just needed to
choose between a specified version in transaction_document_item, or
the version pointed by the header.