Reorganize indexer functions to make API visible.
Reorganize the functions of the indexer module to make the module's
API more visible; i.e. move the functions that define the indexer API
to the top of the namespace, while moving the implementation functions
down to the bottom. We also annotate the two primary functions that
define how other parts of Librarian directly interact with the
indexer; we leave the actual indexer thread as a bare function, since
it shouldn't be called directly.