Make version a computed property of a document.
Instead of providing a slot for the version of a document, make the
version a computed property, as the version of a document should not
be directly changeable, and should be based on the instance's value
instead -- that is, given a document instance, the version of that
instance should be derivable from the instance itself.

We were already doing that inside `write-document!`, calculating the
version and setting it into the `version` slot. This makes it more
explicit.

In the process of disentangling the version, I discovered that due to
how `JSONB` types are persisted by Postgres, the version we calculate
before writing the JSON document would be different than what would be
fetched after reading from the database. This is due to the fact that
Postgres sorts keys in the process of creating the `JSONB` type, so
the following map:

   {
      "foo": 1,
      "bar": 2
   }

would be effectively stored in Postgres as:

   {
      "bar": 2,
      "foo": 1
   }

and after writing the former, retrieving the `JSONB` value would yield
the latter.

Keys in Clojure maps are unordered, but the current implementation
persists key order: converting the above into Clojure maps yields two
equivalent maps, but iterating through each yields different item
orders, and Cheshire uses the iteration order to determine the order
of keys in JSON -- hence, when calculating the SHA-256 digest for the
JSON string, we end up with different digests, even for effectively
equal maps.

So, to make the version calculation deterministic, we enforce sort
order in maps, by providing a multimethod that yields a "canonical"
representation, substituting sorted maps in place of regular unordered
maps. The order of vectors/lists are kept preserved, as usual, and
only maps are canonicalized.

An alternative to the above, of course, is to simply store the JSON
string we generate for the version calculation as-is in the database
as an additinal column of type `JSON`; however, this would be terrible
in terms of disk space, adding to our overhead.