Add store functions for getting a document in a given transaction.
Add store functions for fetching the current version of a document
within a transaction, given any previous writes. This means that the
value returned by e.g. `get-document-by-id-in-transaction` may be
different from `get-document-by-id`, if that document was written to
by the transaction -- the call to `get-document-by-id-in-transaction`
will yield the last written value in the transaction.