Fix storage GC to properly handle dangling refs.
The previous SQL query we were executing was too simplistic, and
didn't properly handle the version reference chain from transaction
roots, and tried to delete inner versions of the chain, as these
weren't referenced directly by transactions.

Modify the query to properly account for the chain, and only delete
document versions that are not reachable from any of the transaction
roots.

To allow deletes to happen, also alter the foreign key on the document
index to allow deletes to cascade to the document index.