Do not cache nil values.
Previously, nil values would be cached under the cache key; this would
mean that, for instance, doing a lookup for the existence of a header
would incorrectly yield nil on the second call, even if the header was
created after the nil value on the first call.

As well, remove the unnecessary `dosync`.