Fix migrations on test start.
This was slightly broken, as some components that were started on
`mount/start` depended on the schema already existing; on a fresh
developer system, this wouldn't get us to bootstrap the test database
by running `lein test`, as `mount/start` wouldn't complete. Succeeding
invocations would run newer migrations.

So, simplify instead and have the migrations be run inside their own
mount context, depending only on the DB data source and migration
config; boot the rest of the system afterwards for the tests.