Replace version representation from hex to base64.
Replace our representation of document versions from hex strings to
base64 strings. Admittedly, the most compact representation would be
the raw 32 bytes of the SHA-256 digest, but that wouldn't be
representable in URLs or parameters; base64 is a decent compromise
between being able to use the version string in URLs and parameters,
and compactness. It might be best to store the version as the raw byte
values (or at least as 4 64-bit values).